requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
websockets==12.0
websockets==12.0
